ensure initial resize is sent on session attach
This updates the client to send an initial resize_pty message immediately after attaching to a session.
It adds a new attached action to ServerAction to trigger this behavior.
Additionally, it adds a terminal query on startup and updates a test in server.zig.

Amp-Thread-ID: https://ampcode.com/threads/T-c3277026-d886-46c3-aae2-11b96a11f74f

+++ src/server.zig
@@ -1728,7 +1728,7 @@ test "buildRedrawMessage" {
     };
     try rows_data.append(testing.allocator, .{ .y = 0, .cells = cells });
 
-    var styles = std.AutoHashMap(u16, redraw.UIEvent.Style.Attributes).init(testing.allocator);
+    var styles = std.AutoHashMap(u32, redraw.UIEvent.Style.Attributes).init(testing.allocator);
     defer styles.deinit();
     try styles.put(1, .{ .fg = 0xFF0000 });
